removed test_u and test_h and added it to test_shared.py
